+++ rusk/src/lib/http/rusk.rs
@@ -7,16 +7,14 @@
 use super::*;
 
 use anyhow::anyhow;
-use dusk_bytes::{DeserializableSlice, Serializable};
+use dusk_bytes::Serializable;
 use dusk_core::abi::ContractId;
-use dusk_core::signatures::bls::PublicKey as BlsPublicKey;
 use dusk_core::stake::{StakeFundOwner, STAKE_CONTRACT};
 use dusk_core::transfer::TRANSFER_CONTRACT;
 use dusk_data_driver::ConvertibleContract;
 use event::RequestData;
 use rusk_profile::CRS_17_HASH;
 use serde::Serialize;
-use serde_json::json;
 use std::sync::mpsc;
 use std::thread;
 
@@ -32,7 +30,6 @@ impl HandleRequest for Rusk {
             ("contracts", Some(_), _) => true,
             ("driver", Some(_), _) => true,
             ("node", _, "provisioners") => true,
-            ("account", Some(_), "status") => true,
             ("node", _, "crs") => true,
             _ => false,
         }
@@ -57,7 +54,6 @@ impl HandleRequest for Rusk {
             }
             ("node", _, "provisioners") => self.get_provisioners(),
 
-            ("account", Some(pk), "status") => self.get_account(pk),
             ("node", _, "crs") => self.get_crs(),
             _ => Err(anyhow::anyhow!("Unsupported")),
         }
@@ -226,24 +222,6 @@ impl Rusk {
             .collect();
 
         Ok(ResponseData::new(serde_json::to_value(prov)?))
-    }
-
-    fn get_account(&self, pk: &str) -> anyhow::Result<ResponseData> {
-        let pk = bs58::decode(pk)
-            .into_vec()
-            .map_err(|_| anyhow::anyhow!("Invalid bs58 account"))?;
-        let pk = BlsPublicKey::from_slice(&pk)
-            .map_err(|_| anyhow::anyhow!("Invalid bls account"))?;
-        let account = self
-            .account(&pk)
-            .map(|account| {
-                json!({
-                    "balance": account.balance,
-                    "nonce": account.nonce,
-                })
-            })
-            .map_err(|e| anyhow::anyhow!("Cannot query the state {e:?}"))?;
-        Ok(ResponseData::new(account))
     }
 
     fn get_crs(&self) -> anyhow::Result<ResponseData> {