+++ rusk/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support for `TransactionData::Blob`
+- Add `mempool_nonce` field to `/on/account/status` response
 
 ## [1.3.0] - 2025-04-17